requirements.txt
python-dotenv
openai
streamlit
webdriver-manager
selenium
beautifulsoup4
undetected-chromedriver